backend/pyproject.toml
[project]
name = "cartsync-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]",
    "aiosqlite",
    "python-jose[cryptography]",
    "argon2-cffi",
    "python-dotenv",
    "redis[asyncio]",
    "sse-starlette",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]
